requirements.txt
scikit-learn
pandas
numpy
faiss-cpu
streamlit
transformers
torch
tqdm
